Move DygraphToStatic Unit Tests into Its Folder
test=develop
